Hak akses website

+++ app/Config/Routes.php
@@ -40,35 +40,35 @@ $routes->get('/dashboard', 'Dashboard::index');
 $routes->post('/dashboard/table', 'Dashboard::showTable');
 
 
-$routes->get('/satuan', 'Unit::index');
-$routes->post('/satuan/table', 'Unit::showTable');
-$routes->post('/satuan/modal', 'Unit::showModal');
-$routes->post('/satuan/save', 'Unit::save');
-$routes->post('/satuan/confirm', 'Unit::confirmDelete');
-$routes->post('/satuan/delete', 'Unit::deleteUnit');
+$routes->get('/satuan', 'Unit::index', ['filter' => 'role:Super Admin, Gudang']);
+$routes->post('/satuan/table', 'Unit::showTable', ['filter' => 'role:Super Admin, Gudang']);
+$routes->post('/satuan/modal', 'Unit::showModal', ['filter' => 'role:Super Admin, Gudang']);
+$routes->post('/satuan/save', 'Unit::save', ['filter' => 'role:Super Admin, Gudang']);
+$routes->post('/satuan/confirm', 'Unit::confirmDelete', ['filter' => 'role:Super Admin, Gudang']);
+$routes->post('/satuan/delete', 'Unit::deleteUnit', ['filter' => 'role:Super Admin, Gudang']);
 
 $routes->get('/barang', 'Item::index');
 $routes->post('/barang/table', 'Item::showTable');
-$routes->post('/barang/modal', 'Item::showModal');
-$routes->post('/barang/save', 'Item::save');
-$routes->post('/barang/confirm', 'Item::confirmDelete');
-$routes->post('/barang/delete', 'Item::deleteItem');
+$routes->post('/barang/modal', 'Item::showModal', ['filter' => 'role:Super Admin, Gudang']);
+$routes->post('/barang/save', 'Item::save', ['filter' => 'role:Super Admin, Gudang']);
+$routes->post('/barang/confirm', 'Item::confirmDelete', ['filter' => 'role:Super Admin, Gudang']);
+$routes->post('/barang/delete', 'Item::deleteItem', ['filter' => 'role:Super Admin, Gudang']);
 
-$routes->get('/stock/tambah', 'Item::showModalAddStock');
+$routes->get('/stock/tambah', 'Item::showModalAddStock', ['filter' => 'role:Super Admin, Gudang']);
 $routes->get('/stock/kurang', 'Item::showModalMinusStock');
 $routes->post('/stock/current', 'Item::showCurrentStock');
-$routes->post('/stock/add', 'Item::addStock');
+$routes->post('/stock/add', 'Item::addStock', ['filter' => 'role:Super Admin, Gudang']);
 $routes->post('/stock/minus', 'Item::minusStock');
 
 $routes->get('/laporan', 'Report::index');
 $routes->post('/laporan/table', 'Report::showTable');
-$routes->post('/laporan/export', 'Report::exportToExcel');
+$routes->post('/laporan/export', 'Report::exportToExcel', ['filter' => 'role:Super Admin, Gudang']);
 
-$routes->get('/pengguna', 'User::index');
-$routes->post('/pengguna/table', 'User::showTable');
-$routes->post('/pengguna/save', 'User::save');
-$routes->post('/pengguna/confirm', 'User::confirmDelete');
-$routes->post('/pengguna/delete', 'User::deleteUser');
+$routes->get('/pengguna', 'User::index', ['filter' => 'role:Super Admin']);
+$routes->post('/pengguna/table', 'User::showTable', ['filter' => 'role:Super Admin']);
+$routes->post('/pengguna/save', 'User::save', ['filter' => 'role:Super Admin']);
+$routes->post('/pengguna/confirm', 'User::confirmDelete', ['filter' => 'role:Super Admin']);
+$routes->post('/pengguna/delete', 'User::deleteUser', ['filter' => 'role:Super Admin']);
 
 $routes->get('/profile', 'User::myProfile');
 $routes->get('/profile/setting', 'User::setMyProfile');
